btrfs: carve out btrfs_buffered_iomap_begin from write path
Carve out reservation and locking from bufferd write sequence in
btrfs_buffered_iomap_begin(). The function performs space reservations
and locks the extent as required.

The writesize is limited to 2M based on a 4K pagesize: 512 pointers in a
page of 4K pages. I am not stuck on this number, but I think it should
be limited so as not to extent lock a huge range.

Anyone else besides me feel the need to change variable name
"write_bytes" to "len"?

Shortcomings:
page handling with respect to nrptrs is messed up. However, this should
be fixed later with switch to iomap.
nowait will be correctly assigned in a later patch.

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.com>