btrfs: define iomap_read_folio_ops
btrfs_read_submit_io() sets all pages in the bio to extent mapped and
submit the bio.
Use btrfs_bioset so iomap allocates btrfs_bio for bios to be submitted
for btrfs.

For compressed/encoded reads, bbio->file_offset needs to be set. This is
extracted from iomap.offset since one bbio is used for one iomap setup
and the entire extent must be read (no offsets). So, a simple assignment
of bbio.file_offset = iomap.offset is good enough.

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.com>