iomap: Introduce IOMAP_ENCODED
An encoded extent must be read completely and cannot be partial.
If the filesystem returns iomap.type as IOMAP_ENCODED, create a
bio to read the entire scope of the iomap.

If required, the bio consumes the folios set in readahead_control so
as to cover the entire extent.

The submit_io() function is changed to accomodate because btrfs needs
to fill in the file_offset for the btrfs_bio

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.com>