iomap: Introduce IOMAP_ENCODED
An encoded extent must be read completely and cannot be partial.
If the filesystem returns iomap.type as IOMAP_ENCODED, create a
bio to read the entire scope of the iomap.

If required, the bio consumes the folios set in readahead_control so
as to cover the entire extent.

While processing the iterator, reset the iterator to iomap's offsets
because we would need to process the entire extent even if we are
reading a part of it.

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.com>