btrfs: add WARN_ON() on incorrect lock range
Add a WARN_ON(start > end) to make sure that the locking happens on the
correct range and no incorrect nodes (with state->start > state->end)
are added to the tree.

Note, an example when this situation happened was truncating a file to
the current file size (with the extent page locking interchange
patches applied).

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.com>