btrfs: return a max of BTRFS_MAX_COMPRESSED for async writebacks
For compressed writebacks, a compressed extent can be a maximum size of
128K. This is done so we don't have to break the bio into chunks as done
in cow_file_range_async(). Each bio can be dealt as a separate writeback
thread, and will get an struct async_extent to work with.

Also, skip disk allocation done through btrfs_run_delalloc_range(). The
disk allocation is performed after the data is compressed.

Set the limit of the compressed extent in inode.c:btrfs_map_blocks().
The blocks are mapped after the compression, so we will have to provide
an iomap with 0. This should be safe as long as for async writebacks we
either:
 - transform the bio to compressed bio and writeback the compressed bio.
   The compressed bio has it's own compressed extent created which is the
   true sector it will be written to.
 - submit uncompressed bio. In this case a cow_file_range() is called
   and the bio's sector is re-written.

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.com>