btrfs: lzo compression to work with bio
Logical new section: async compression writes

Design for writeback is pretty much similar to how compressed reads are
handled.
Instead of btrfs handling the page lookup, lock and writeback; this is
routed through iomap. iomap collects the pages and provides btrfs with a
bio. btrfs compresses the pages on the bio, allocates a new
compressed_bio and submits the writes. Finally once write is complete,
call bio_endio() on the original bio.

Add lzo_compress_bio() which compresses a bio (provided by iomap), and
puts them in pages. The total compressed length is returned by the
function, or a negative number for error.

As a temporary measure for the switch, I have added
compression_compress_bio() function.

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.com>