btrfs: Use bio with zstd
Accept a bio as a function parameter, compress the pages and return the
length of the compressed size. The loop is converted to iterate through
the pages of the bio. If out page is exhausted, a new one is created and
compression is attempted on the rest of the input folio.

Note, this breaks zstd until compression_compress_bio() is incorporated.

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.com>