btrfs: zlib compression to use bio
Same as zstd.
Accept a bio as a function parameter, compress the pages and return the
length of the compressed size. The loop is converted to iterate through
the pages of the bio. If out page is exhausted, a new one is created and
compression is attempted on the rest of the input folio.

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.com>