btrfs: fold async_chunk into async_extent
This is a preparatory patch to make compressed extents based on bios.

As compression is performed on bios, we know the pages are contiguous
and it is less than or equal to BTRFS_MAX_COMPRESSED.

So, there is no need of async_chunk. Remove it and fold the fields into
async_extent.

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.com>