btrfs: best effort extent locking
Introduce lock_extent_best_effort(). Lock extents with best effort.
The function attempts to lock the entire range submitted. However, if
the locking falls short because part of the extent is already locked,
the function returns the bits already locked, provided it starts from
the same offset and has a minimum of the size passed.
Otherwise, it waits for min_size to be unlocked starting from the start
variable passed.

This function is useful while doing writebacks to proceed, even if the
entire range is not available for locking.

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.com>