btrfs: btrfs_prepare_ioend() to prepare btrfs_bio
btrfs_prepare_ioend() prepares the btrfs_bio for submission. This
includes initializing the bbio, setting the inode and the file offset
and setting up the ordered_extent associated with the bbio.

The bbio is setup before returning the status (in case of error), so
that bio_endio() in iomap_submit_ioend() executes
btrfs_writepages_endio() as opposed to iomap_finish_ioend().

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.com>