btrfs: lock extents during writes unconditionally
This is to fix a race between buffered write() and writepages(), with
respect to inode size.

Extents are not locked during writes if the write is beyond the end of
file. However, if a write is going on simultaneously with a writeback,
it will increase the inode size. This will cause some extents not to be
marked EXTENT_DELALLOC because iomap_end() is not called as yet and the
writeback will not find the extent to write even though the inode size
has increased.

Lock the extents during write so writeback cannot initiate until the
write has completed.

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.com>