btrfs: cow_file_range() to return ordered extent
iomap works on offset and calls back map_blocks() as and when required.
Instead of allocating all the ordered extents, allocate the extents as
required by the iomap code.

In order to do this, change the allocation sequence to return the
ordered extent. This obviates the need to search for extent map
later on since an ordered extent has all the fields to fill in the
iomap.

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.com>