btrfs: use ordered extent to fill iomap fields
Create only one ordered extent at offset, as opposed to the entire
range. Once btrfs_run_delalloc_range() returns an ordered extent, use it
to fill the iomap fields.

This is done to remove the dependency on getting an extent map.
This is a cleaner process than creating the ordered extent for the
entire range, because range extents needs to be re-searched and it
messes up how we complete our ordered extents.

As in the previous patch, range end is just advisory and not mandatory.

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.com>