btrfs: Introduce btrfs_writeback to lock extents during writeback
Locking of extents is performed for the entire duration of the
writeback, starting from btrfs_writepages() is called until the bio
generated by iomap code finishes.

This is primarily done for asynchronous I/O. The problem is ordered
extent allocation is performed outside of and after pages are allotted
for bio and set for writeback. Another process can race against this and
lock the extent after writepages() extent lock is release but before
lock extent for allocation of compressed bio. The racing processes
wait on writeback on a page to complete, but that does not happen
because async writeback has not completed. This structure helps lock
for the entire duration of the writeback.

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.com>